# openapi.json
{
  "components": {
    "schemas": {
      "CommandName": {
        "description": "Batchalign command name (e.g. `\"morphotag\"`, `\"align\"`).",
        "type": "string"
      },
...
      "DurationSeconds": {
        "description": "Duration measured in fractional seconds.",
        "format": "double",
        "type": "number"
      },
      "ErrorResponse": {
        "description": "Standard error response body, matching FastAPI's `HTTPException` format.\n\nAll non-2xx responses from the server use this shape so that clients\ncan parse errors uniformly. The one exception is\n[`ServerError::JobConflict`](crate::error::ServerError::JobConflict),\nwhich nests a `conflicts` array inside `detail`.",
        "properties": {
          "detail": {
            "description": "Human-readable description of what went wrong. For validation\nerrors this includes the specific field or constraint that failed.",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
      "FileResult": {
        "description": "Result for a single processed file.",
        "properties": {
          "content": {
            "description": "Processed file content (CHAT text or CSV).  Empty string when\n`error` is `Some`.",
            "type": "string"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType",
            "description": "MIME-like content discriminator: `\"chat\"` for CHAT files (default),\n`\"csv\"` for tabular output (e.g. opensmile features)."
          },
          "error": {
            "description": "Human-readable error message if this file failed processing.\n`None` for successfully processed files.",
            "type": [
              "string",
              "null"
            ]
          },
          "filename": {
            "$ref": "#/components/schemas/FileName",
            "description": "Original filename from the submission (e.g. \"01DM_18.cha\")."
          }
        },
        "required": [
          "filename"
        ],
        "type": "object"
      },
      "FileStatusEntry": {
        "description": "Per-file status within a job.\n\nTracks processing state, timing, progress, and error details for a\nsingle file.  The server updates these entries as workers report\nresults, and they are included in `JobInfo.file_statuses`.",
        "properties": {
          "bug_report_id": {
            "description": "Opaque identifier linking to a persisted bug report (for post-mortem\nanalysis).  Generated by the server when a file fails unexpectedly.",
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "description": "Human-readable error message.  Present only when `status` is `Error`.",
            "type": [
              "string",
              "null"
            ]
          },
          "error_category": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FailureCategory",
                "description": "Broad classification of the failure. Helps the dashboard and retry\nlogic group failures without parsing free-form messages."
              }
            ]
          },
          "error_codes": {
            "description": "CHAT validation error codes (e.g. `[\"E362\", \"E704\"]`) when the failure\nwas a validation rejection.  Empty or absent for non-validation errors.",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "error_line": {
            "description": "1-based line number in the CHAT file where the first error was\ndetected, if applicable.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "filename": {
            "$ref": "#/components/schemas/FileName",
            "description": "Original filename from the submission (e.g. \"01DM_18.cha\")."
          },
          "finished_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnixTimestamp",
                "description": "Unix timestamp (seconds since epoch) when processing finished\n(successfully or with error)."
              }
            ]
          },
          "next_eligible_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnixTimestamp",
                "description": "When the file is deferred for a retry, the earliest unix timestamp at\nwhich another attempt should start. `None` when no retry is pending."
              }
            ]
          },
          "progress_current": {
            "description": "Number of sub-steps completed so far (e.g. utterances processed).\nUsed with `progress_total` to drive progress bars.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "progress_label": {
            "description": "Human-readable label for the current processing stage\n(e.g. \"Morphosyntax\", \"Forced Alignment\").\n\nThis is derived from `progress_stage` by the server so operators can\nrender friendly text without hard-coding label mappings in every UI.",
            "type": [
              "string",
              "null"
            ]
          },
          "progress_stage": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FileProgressStage",
                "description": "Machine-readable code for the current in-flight processing stage.\n\nThis is the stable field clients should use for branching or\nconditional UI. [`Self::progress_label`] is derived display text."
              }
            ]
          },
          "progress_total": {
            "description": "Total number of sub-steps expected for this file.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "started_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnixTimestamp",
                "description": "Unix timestamp (seconds since epoch) when the worker began processing\nthis file."
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/FileStatusKind",
            "description": "Current lifecycle state of this file."
          }
        },
        "required": [
          "filename",
          "status"
        ],
        "type": "object"
      },
...
      "JobId": {
        "description": "Server-assigned UUID (v4) for a job.",
        "type": "string"
      },
      "JobInfo": {
        "description": "`GET /jobs/{id}` response — job progress.\n\nContains full detail about a single job, including per-file statuses.\nPolled by the CLI to drive progress bars and by the dashboard for\nlive-updating tables.",
        "properties": {
          "active_lease": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LeaseRecord",
                "description": "Active lease information when this job is currently claimed by a node."
              }
            ]
          },
          "command": {
            "$ref": "#/components/schemas/CommandName",
            "description": "Batchalign command that was submitted (e.g. \"morphotag\", \"align\")."
          },
          "completed_at": {
            "description": "ISO 8601 timestamp of when the job reached a terminal state.",
            "type": [
              "string",
              "null"
            ]
          },
          "completed_files": {
            "description": "Number of files that have finished processing (status `Done`).\nInvariant: `0 <= completed_files <= total_files`.",
            "format": "int64",
            "type": "integer"
          },
          "current_file": {
            "description": "Filename currently being processed.  `None` when the job is queued\nor has reached a terminal state.",
            "type": [
              "string",
              "null"
            ]
          },
          "duration_s": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DurationSeconds",
                "description": "Wall-clock duration from first file dispatched to last file finished,\nin seconds.  `None` while the job is still active."
              }
            ]
          },
          "error": {
            "description": "Job-level error message (e.g. worker pool exhaustion, memory gate\ntimeout).  Distinct from per-file errors in `file_statuses`.",
            "type": [
              "string",
              "null"
            ]
          },
          "file_statuses": {
            "description": "Per-file processing status, one entry per submitted file.",
            "items": {
              "$ref": "#/components/schemas/FileStatusEntry"
            },
            "type": "array"
          },
          "job_id": {
            "$ref": "#/components/schemas/JobId",
            "description": "Server-assigned UUID (v4) for this job."
          },
          "lang": {
            "$ref": "#/components/schemas/LanguageCode3",
            "description": "3-letter ISO language code.  Defaults to `\"eng\"` when absent in JSON."
          },
          "next_eligible_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnixTimestamp",
                "description": "When the job is deferred before execution, the earliest unix timestamp\nat which it should be retried. `None` when the job is immediately\neligible to run."
              }
            ]
          },
          "num_workers": {
            "description": "Number of concurrent Python workers used for this job.  Determined\nby the pool at dispatch time.  `None` for queued jobs.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "options": {
            "description": "Typed submitted command options captured at job creation time.\n\nThe dashboard uses this to show the original argument/config payload for\ndebugging and operator review without reconstructing a lossy CLI string."
          },
          "source_dir": {
            "description": "Client's original input directory path, used for display in the\ndashboard and CLI output.  May be empty for content-mode submissions.",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Current lifecycle state of the job."
          },
          "submitted_at": {
            "description": "ISO 8601 timestamp of when the job was submitted (e.g.\n`\"2026-01-15T10:00:00Z\"`).",
            "type": [
              "string",
              "null"
            ]
          },
          "submitted_by": {
            "description": "IP address or Tailscale identifier of the submitting client.\nUsed for conflict detection (same `submitted_by` + filename =\nduplicate).",
            "type": [
              "string",
              "null"
            ]
          },
          "submitted_by_name": {
            "description": "Human-readable hostname of the submitting machine, resolved from\nTailscale or reverse DNS.  For dashboard display only.",
            "type": [
              "string",
              "null"
            ]
          },
          "total_files": {
            "description": "Total number of files in this job (immutable after submission).",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "status",
          "command",
          "options",
          "total_files",
          "completed_files"
        ],
        "type": "object"
      },
...
      "JobResultResponse": {
        "description": "`GET /jobs/{id}/results` response — completed job results.",
        "properties": {
          "files": {
            "description": "Per-file results (empty until the job completes).",
            "items": {
              "$ref": "#/components/schemas/FileResult"
            },
            "type": "array"
          },
          "job_id": {
            "$ref": "#/components/schemas/JobId",
            "description": "Unique job identifier."
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Terminal status of the job."
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "type": "object"
      },
...
      "LanguageCode3": {
        "description": "3-letter ISO 639-3 language code (e.g. `\"eng\"`, `\"spa\"`).",
        "type": "string"
      },
      "LeaseRecord": {
        "description": "Lease metadata for a claimed schedulable unit.\n\nIn single-node mode this is coarse-grained and attached at the job level:\nthe local dispatcher claims a queued job for a node, records the lease, and\nlater clears it when the runner exits. Fleet mode will extend the same shape\nwith real cross-node renewal and expiry handling.",
        "properties": {
          "expires_at": {
            "$ref": "#/components/schemas/UnixTimestamp",
            "description": "Unix timestamp when the lease should be considered expired if not renewed."
          },
          "heartbeat_at": {
            "$ref": "#/components/schemas/UnixTimestamp",
            "description": "Unix timestamp when the lease was created or most recently renewed."
          },
          "leased_by_node": {
            "$ref": "#/components/schemas/NodeId",
            "description": "Identifier of the node that currently owns the lease."
          }
        },
        "required": [
          "leased_by_node",
          "heartbeat_at",
          "expires_at"
        ],
        "type": "object"
      },
...
      "StatusMessageResponse": {
        "description": "Response body for mutating operations that return a status confirmation\n(e.g. cancel, delete, restart).\n\nReturned with HTTP 200 on success. Clients can branch on `status` for\nprogrammatic handling and display `message` to users.",
        "properties": {
          "message": {
            "description": "Human-readable sentence describing the action taken (e.g.\n`\"Job abc123 has been cancelled.\"`).",
            "type": "string"
          },
          "status": {
            "description": "Machine-readable outcome token (e.g. `\"cancelled\"`, `\"deleted\"`,\n`\"restarted\"`). Suitable for switch/match in client code.",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "type": "object"
      },
      "UnixTimestamp": {
        "description": "Unix timestamp as fractional seconds since epoch.",
        "format": "double",
        "type": "number"
      },
...
    }
  },
  "info": {
    "description": "batchalign3 application: types, server, worker pool, cache",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "batchalign-app",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/bug-reports": {
      "get": {
        "description": "Bug reports are JSON files written to disk by the validation layer when\nit detects semantic errors (e.g., alignment mismatches, monotonicity\nviolations). This endpoint scans the reports directory by mtime so the\ndashboard can surface recent failures without requiring database queries.\nMissing or unreadable files are silently skipped.",
        "operationId": "list_bug_reports",
        "parameters": [
          {
            "description": "Maximum number of reports to return",
            "in": "query",
            "name": "limit",
            "required": true,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bug report documents (newest first)"
          }
        },
        "summary": "Return the most recent bug reports, sorted newest-first.",
        "tags": [
          "bug-reports"
        ]
      }
    },
    "/bug-reports/{id}": {
      "get": {
        "description": "The ID corresponds to the JSON filename on disk without the `.json`\nextension. Returns the raw JSON document as-is so the dashboard can\nrender full diagnostic detail (stack traces, file paths, expected vs.\nactual values).",
        "operationId": "get_bug_report",
        "parameters": [
          {
            "description": "Bug report ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bug report JSON document"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bug report not found"
          }
        },
        "summary": "Return a single bug report by its ID (filename stem).",
        "tags": [
          "bug-reports"
        ]
      }
    },
...
    "/jobs/{job_id}": {
      "delete": {
        "description": "Returns 409 if the job is still running -- the caller must cancel it first.\nDeleting a job removes it from the in-memory store and SQLite, and broadcasts\na `JobDeleted` event to connected WebSocket/SSE clients.",
        "operationId": "delete_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessageResponse"
                }
              }
            },
            "description": "Job deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job still running"
          }
        },
        "summary": "Permanently remove a terminal job and its associated state.",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "description": "This is the primary polling endpoint for the CLI: the client calls it\nrepeatedly until `status` reaches a terminal value.",
        "operationId": "get_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            },
            "description": "Job detail"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Return full detail for a single job, including per-file statuses.",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/jobs/{job_id}/results": {
      "get": {
        "description": "Blocks until the job has completed, failed, or been interrupted -- returns\n409 for jobs that are still running or queued. In content mode the response\nincludes the full CHAT text of each output file; in paths mode only filenames\nare returned (the files were written directly to disk by the runner).\nCancelled jobs are intentionally excluded because they may have incomplete\nresults.",
        "operationId": "get_results",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResultResponse"
                }
              }
            },
            "description": "Results for terminal jobs"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Retrieve all output files for a terminal job.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/results/{filename}": {
      "get": {
        "description": "Waits for the individual file to reach a terminal status, not the whole job,\nso clients can fetch results incrementally as files finish. Supports\nstem-based matching (e.g., requesting `sample.wav` will find `sample.cha`)\nto handle the common case where the output extension differs from the input.",
        "operationId": "get_single_result",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Requested output filename",
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResult"
                }
              }
            },
            "description": "One file result"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "File not found"
          }
        },
        "summary": "Retrieve the output for a single file within a job.",
        "tags": [
          "jobs"
        ]
      }
    },
...
  },
...
}